pyproject.toml
[project]
name = "kinda-jarvis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.6.15",
    "dotenv>=0.9.9",
    "google-genai>=1.31.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "joblib==1.5.1",
    "numpy==2.3.1",
    "pandas==2.3.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "scikit-learn==1.7.0",
    "scipy==1.16.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "threadpoolctl==3.6.0",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
]
